Moves the zip_create_from_directory_preserve_symlinks_with_options function into a separate module and adds appropriate documentation
