+++ src/lib.rs
@@ -6,9 +6,10 @@ pub use crate::write::*;
 pub mod default_entry_handler;
 pub mod entry_handler;
 mod file_utils;
-pub mod preserver_symlinks_handler;
+pub mod preserve_symlinks_handler;
 pub mod read;
 pub mod write;
+pub mod preserve_symlinks;
 
 #[cfg(test)]
 mod tests {
